+++ HISTORY.md
@@ -6,6 +6,7 @@
 3. rename data_store
 4. Move edit IDE colours
 5. Collapse home/jeff to ~ on save project
+6. refactor get_versions into data_store an key on project.name
 
 ## Version 1.0.10 - 16 Jul 2026
 1. Add copy toenv and copy to live button to compare